Free raffle tickets that entered no draw, pet XP that never saved, cooking UI
RAFFLE. The pets screen advertises fortune as "Free weekly-boss raffle entries
each day - real odds to win". FREE is the promise. But the draw builds its pool
from boss_hit with HAVING SUM(damage) > 0, so a member who never swung at the
boss was not in the pool at all and their tickets entered nothing. The stat paid
out only for people already fighting, who had damage tickets anyway. The RAFFLE
pool alone now includes anyone holding fortune; `pool` still means "everyone who
fought" and keeps driving participation XP, spin tokens, badges, seeds and reward
items, none of which are free and none of which should go to someone absent.

PET XP. pet-level.js ran LEAST($3, $4) with two JS-number params. Postgres has
nothing to infer from, types them TEXT, and the insert dies on the integer xp
column with 42804. farm.js already carries ::int casts; this copy never got them,
so every pet-XP grant through here failed - and failed INVISIBLY, because the
.catch falls back to a computed value and returns ok:true with a plausible new
level while nothing was written. Swept the codebase; no other uncast LEAST.

WHERE THE MONEY IS. Store credit removed per owner's call. It is a real claim but
members redeem it in merchandise at retail, not cash, so subtracting it from the
cash position overstated what the shop owes in money.

COOKING. Upgrades were the last of four stacked cards, below a 64-row recipe
list, so the thing you spend gold on was the thing nobody scrolled to - now
top-level Recipes / Pantry / Upgrades tabs.

The timing bar is the forge's: same band distances (0.022/0.055/0.10/0.16),
where cooking had been running at exactly double, so 72% of the bar scored and it
never felt like a skill check. Its tightest tier had no band drawn at all, so the
target was invisible - all four are now drawn, at widths that ARE the scoring
numbers doubled, and a spoon rides the marker like the forge's hammer.

The reveal is portalled to <body>. It is a direct child of `.stack.reveal`, and
`.reveal > *` applies fade-in-up with fill-mode both, which leaves a permanent
transform on the element. A transform other than none makes an element the
containing block for its own position:fixed children, so inset:0 stopped meaning
the viewport and the payoff card laid itself out somewhere down a very long page.

+++ src/lib/marketplace/pet-level.js
@@ -96,10 +96,18 @@ export async function addEquippedPetXp(buyerId, amount) {
     const prevLevel = petLevelForXp(before?.xp || 0, rarity);
     const row = await db
         .queryOne(
+            // The ::int casts are REQUIRED. LEAST($3, $4) with two JS-number params gives Postgres nothing to
+            // infer from, so it types them TEXT and the insert dies on the integer xp column with 42804 —
+            // "column xp is of type integer but expression is of type text". farm.js already carries the casts;
+            // this copy never got them, so every pet-XP grant that came through here failed.
+            //
+            // The failure was invisible because the .catch below falls back to a COMPUTED xp, so the caller was
+            // told ok:true with a plausible new level while nothing was written. Pets appeared to gain XP and
+            // silently reset on the next read.
             `INSERT INTO mkt_pet_level (buyer_id, pet_id, xp, last_tick_at, updated_at)
-             VALUES ($1, $2, LEAST($3, $4), NOW(), NOW())
+             VALUES ($1, $2, LEAST($3::int, $4::int), NOW(), NOW())
              ON CONFLICT (buyer_id, pet_id)
-             DO UPDATE SET xp = LEAST(mkt_pet_level.xp + $3, $4), updated_at = NOW()
+             DO UPDATE SET xp = LEAST(mkt_pet_level.xp + $3::int, $4::int), updated_at = NOW()
              RETURNING xp`,
             [buyerId, petId, add, maxXp]
         )
